refactor: move spyFunc and stubFunc to it's own folder.
Make it look like other plugins.
